Consolidate CDataType into api/types.rs and revert error.rs to api/
Move CDataType, constants (SQL_NULL_DATA, SQL_NO_TOTAL, C_TYPES_EXTENDED),
and type aliases from the standalone cdata_types.rs into api/types.rs,
co-locating all ODBC type definitions. Move error.rs back to api/error.rs
since CDataType can now directly use OdbcError from its sibling module.

Made-with: Cursor
