+++ odbc/src/conversion/binary.rs
@@ -4,8 +4,8 @@ use arrow::array::{Array, GenericByteArray};
 use arrow::datatypes::GenericBinaryType;
 use serde_json::Value;
 
+use crate::api::CDataType;
 use crate::api::ParameterBinding;
-use crate::cdata_types::CDataType;
 use crate::conversion::error::JsonBindingError;
 use crate::conversion::error::{ReadArrowError, UnsupportedOdbcTypeSnafu, WriteOdbcError};
 use crate::conversion::param_binding::buffer_data_len;
