+++ odbc/src/conversion/date.rs
@@ -4,8 +4,8 @@ use chrono::{Datelike, NaiveDate};
 use odbc_sys as sql;
 use serde_json::Value;
 
+use crate::api::CDataType;
 use crate::api::ParameterBinding;
-use crate::cdata_types::CDataType;
 use crate::conversion::error::{JsonBindingError, UnsupportedCDataTypeSnafu};
 use crate::conversion::error::{ReadArrowError, UnsupportedOdbcTypeSnafu, WriteOdbcError};
 use crate::conversion::param_binding::{read_char_str, read_unaligned, read_wchar_str};
