+++ odbc/src/conversion/number.rs
@@ -2,8 +2,8 @@ use arrow::array::{Array, ArrowPrimitiveType, PrimitiveArray};
 use odbc_sys as sql;
 use serde_json::Value;
 
+use crate::api::CDataType;
 use crate::api::ParameterBinding;
-use crate::cdata_types::CDataType;
 use crate::conversion::error::{
     IntervalFieldOverflowSnafu, NumericValueOutOfRangeSnafu, ReadArrowError,
     UnsupportedOdbcTypeSnafu, WriteOdbcError,
